Complete crab cups

+++ AdventOfCode2020/Day23/CupParser.cs
@@ -5,6 +5,6 @@ namespace AdventOfCode2020.Day23
     public static class CupParser
     {
         public static CupCollection Parse(string input) 
-            => new(input.Select(c => long.Parse(c.ToString())));
+            => new(input.Select(c => int.Parse(c.ToString())));
     }
 }
